refactor(cli): give the launch selection and backend choice their own modules
`exec.rs` was 31 production lines under the 1500-line budget, and the selection
added enough to cross it. `check-file-size` asks for a module tree rather than a
bigger file, so this splits along the two seams the change actually touched:

- `exec/sidecar_selection.rs` — the variant choice and the declared-vs-recorded
  cross-check, with the tests for both. Cohesive on its own terms: it weighs
  what the catalog declared against what the image recorded, and both facts are
  now readable in one place.
- `exec/backend_select.rs` — the pre-existing backend/egress selection cluster,
  moved unchanged and re-exported, so no call site moves. Two of its functions
  were only reachable from their own tests, and those tests came with them.

`check-cli-runtime-surface` exempts `exec.rs` for exactly this selection; the
extracted file gets its own rule-scoped entry naming why it is the same reach
rather than a new one.

Behaviour is unchanged. 12,873 tests pass, Clippy is clean, and all 62 xtask
gates are green.
