+++ crates/mvm-cli/src/exec/backend_select.rs
@@ -1,0 +1,225 @@
+//! Which backend may serve a transient run, given its egress posture.
+//!
+//! Selection and validation are one cluster: the choice is made from the
+//! requested hypervisor, the environment override and the image kind, and it is
+//! only sound if the resulting backend can actually enforce the run's network
+//! policy. Separating the two would let a caller pick without checking.
+//!
+//! Split out of `exec.rs` to keep that file inside its production-line budget;
+//! every entry point is re-exported from there, so call sites are unchanged.
+
+use anyhow::{Context, Result, anyhow};
+use mvm_core::vm_backend::RequiredCapabilities;
+use mvm_runtime::backend::AnyBackend;
+
+pub(crate) fn select_exec_backend(
+    image_requested: bool,
+    network_policy: &mvm_core::network_policy::NetworkPolicy,
+    requested: Option<&str>,
+) -> Result<AnyBackend> {
+    // CLI `--hypervisor` wins over the MVM_HYPERVISOR/MVM_BACKEND env override.
+    let backend_override = requested
+        .and_then(normalize_backend_override)
+        .or_else(explicit_hypervisor_override);
+    let backend_name = select_backend_name_for_egress(
+        backend_override.as_deref(),
+        image_requested,
+        network_policy,
+        "OCI --image runs with outbound egress enabled",
+    )?;
+    AnyBackend::require_hypervisor_selectable(&backend_name)?;
+    Ok(AnyBackend::from_hypervisor(&backend_name))
+}
+
+/// An explicit workload-backend override from the environment. The transient
+/// run path otherwise auto-detects the backend; this lets `MVM_HYPERVISOR`
+/// (or `MVM_BACKEND`) pin it — e.g. `libkrun`, whose vsock-tunnel egress path
+/// the auto-detected default would otherwise never select on this host. Every
+/// `select_exec_backend` call site reads the same value, so the admitted plan's
+/// backend and the boot backend agree.
+fn explicit_hypervisor_override() -> Option<String> {
+    ["MVM_HYPERVISOR", "MVM_BACKEND"]
+        .into_iter()
+        .filter_map(std::env::var_os)
+        .find_map(|raw| normalize_backend_override(&raw.to_string_lossy()))
+}
+
+/// Normalize a backend-override string (trim + lowercase); a blank value yields
+/// `None` so an empty env var is treated as "unset" rather than an invalid
+/// backend name.
+fn normalize_backend_override(raw: &str) -> Option<String> {
+    let value = raw.trim().to_ascii_lowercase();
+    (!value.is_empty()).then_some(value)
+}
+
+pub(crate) fn select_backend_name_for_egress(
+    backend_override: Option<&str>,
+    image_requested: bool,
+    network_policy: &mvm_core::network_policy::NetworkPolicy,
+    workload: &str,
+) -> Result<String> {
+    if let Some(backend_name) = backend_override {
+        validate_backend_for_egress(backend_name, image_requested, network_policy, workload)?;
+        return Ok(backend_name.to_string());
+    }
+
+    if !requires_vsock_proxy_backend(image_requested, network_policy) {
+        return Ok(AnyBackend::auto_select().name().to_string());
+    }
+
+    AnyBackend::select_capable_available(&vsock_proxy_backend_requirements())
+        .map(|backend| backend.name().to_string())
+        .map_err(|e| anyhow!("{workload} require a NIC-less host-vsock-proxy backend: {e}"))
+}
+
+pub(crate) fn validate_backend_for_egress(
+    backend_name: &str,
+    image_requested: bool,
+    network_policy: &mvm_core::network_policy::NetworkPolicy,
+    workload: &str,
+) -> Result<()> {
+    if !requires_vsock_proxy_backend(image_requested, network_policy) {
+        return Ok(());
+    }
+
+    let backend = AnyBackend::from_hypervisor(backend_name);
+    let missing = backend
+        .capabilities()
+        .shortfall(&vsock_proxy_backend_requirements());
+    if missing.is_empty() {
+        let available = backend
+            .is_available()
+            .with_context(|| format!("probing backend {backend_name} availability"))?;
+        if available {
+            return Ok(());
+        }
+        anyhow::bail!(
+            "{workload} require a NIC-less host-vsock-proxy backend; backend {backend_name} is unavailable on this host"
+        );
+    }
+
+    anyhow::bail!(
+        "{workload} require a NIC-less host-vsock-proxy backend; backend {backend_name} lacks [{}]",
+        missing.join(", ")
+    );
+}
+
+fn requires_vsock_proxy_backend(
+    image_requested: bool,
+    network_policy: &mvm_core::network_policy::NetworkPolicy,
+) -> bool {
+    image_requested && network_policy.allows_egress()
+}
+
+fn vsock_proxy_backend_requirements() -> RequiredCapabilities {
+    RequiredCapabilities {
+        vsock: true,
+        no_routable_guest_nic: true,
+        host_vsock_proxy: true,
+        ..Default::default()
+    }
+}
+
+pub(crate) fn validate_image_egress_backend(
+    backend: &AnyBackend,
+    image_requested: bool,
+    network_policy: &mvm_core::network_policy::NetworkPolicy,
+) -> Result<()> {
+    if !image_requested || !network_policy.allows_egress() {
+        return Ok(());
+    }
+    let caps = backend.capabilities();
+    if caps.vsock && caps.no_routable_guest_nic && caps.host_vsock_proxy {
+        return Ok(());
+    }
+    anyhow::bail!(
+        "OCI --image runs with outbound egress enabled require a NIC-less host-vsock-proxy backend; \
+         backend {} does not advertise {{vsock,no_routable_guest_nic,host_vsock_proxy}}",
+        backend.name()
+    );
+}
+
+pub(crate) fn validate_image_egress_backend_name(
+    backend_name: &str,
+    image_requested: bool,
+    network_policy: &mvm_core::network_policy::NetworkPolicy,
+) -> Result<()> {
+    let backend = AnyBackend::from_hypervisor(backend_name);
+    validate_image_egress_backend(&backend, image_requested, network_policy)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use mvm_core::util::test_env::TestEnv;
+
+    #[test]
+    fn validate_backend_for_egress_refuses_unavailable_hvf_before_boot_work() {
+        let _guard = mvm_runtime::base::runtime_meta::HOME_TEST_LOCK
+            .lock()
+            .unwrap_or_else(|e| e.into_inner());
+        let mut env = TestEnv::new();
+        let policy = mvm_core::network_policy::NetworkPolicy::allow_list(vec![
+            mvm_core::network_policy::HostPort::new("example.com", 443),
+        ]);
+        env.set("MVM_HVF_SUPERVISOR_PATH", "/no/such/mvm-hvf-supervisor");
+        let err = validate_backend_for_egress(
+            "hvf",
+            true,
+            &policy,
+            "OCI --image runs with outbound egress enabled",
+        )
+        .expect_err("unavailable hvf must fail closed before OCI work");
+        // SAFETY: test-only env mutation.
+        unsafe {
+            std::env::remove_var("MVM_HVF_SUPERVISOR_PATH");
+        }
+        // HVF always advertises the NIC-less host-vsock-proxy egress caps (they
+        // are unconditional — the fail-closed posture), so the capability
+        // shortfall is empty and the refusal comes from the availability probe:
+        // a host whose supervisor can't launch is unavailable, not egress-capable.
+        let msg = err.to_string();
+        assert!(msg.contains("NIC-less host-vsock-proxy backend"));
+        assert!(msg.contains("backend hvf is unavailable on this host"));
+    }
+
+    #[cfg(all(target_os = "macos", target_arch = "aarch64"))]
+    #[test]
+    fn select_backend_name_for_egress_picks_hvf_when_proxy_support_is_available() {
+        let _guard = mvm_runtime::base::runtime_meta::HOME_TEST_LOCK
+            .lock()
+            .unwrap_or_else(|e| e.into_inner());
+        let mut env = TestEnv::new();
+        let dir = tempfile::tempdir().expect("tempdir");
+        let supervisor = dir.path().join("mvm-hvf-supervisor");
+        std::fs::write(&supervisor, b"stub").expect("stub supervisor");
+        let policy = mvm_core::network_policy::NetworkPolicy::allow_list(vec![
+            mvm_core::network_policy::HostPort::new("example.com", 443),
+        ]);
+
+        env.set("MVM_HVF_SUPERVISOR_PATH", &supervisor);
+        let selected = select_backend_name_for_egress(
+            None,
+            true,
+            &policy,
+            "OCI --image runs with outbound egress enabled",
+        )
+        .expect("hvf should satisfy the proxy backend requirement");
+
+        assert_eq!(selected, "hvf");
+    }
+
+    #[test]
+    fn normalize_backend_override_trims_lowercases_and_drops_blank() {
+        assert_eq!(
+            normalize_backend_override("  LibKrun \n"),
+            Some("libkrun".to_string())
+        );
+        assert_eq!(
+            normalize_backend_override("firecracker"),
+            Some("firecracker".to_string())
+        );
+        assert_eq!(normalize_backend_override("   "), None);
+        assert_eq!(normalize_backend_override(""), None);
+    }
+}
